requirements.txt
discord.py==2.3.2
tweepy==4.14.0
python-dotenv==1.0.0
supabase==2.0.3
schedule==1.2.0
requests==2.31.0
PyNaCl==1.5.0
certifi==2024.2.2 